+++ src/datagen/java/blusunrize/immersiveengineering/data/Recipes.java
@@ -45,6 +45,7 @@ import blusunrize.immersiveengineering.data.resources.RecipeWoods;
 import blusunrize.immersiveengineering.data.resources.SecondaryOutput;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
+import net.minecraft.advancements.CriterionTriggerInstance;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.data.HashCache;
 import net.minecraft.data.recipes.*;
@@ -1195,17 +1196,17 @@ public class Recipes extends RecipeProvider
 				makeIngredient(IETags.clay),
 				makeIngredient(Tags.Items.INGOTS_BRICK),
 				makeIngredient(Tags.Items.SANDSTONE),
-				out);
+				has(IETags.clay), out);
 		addCornerStraightMiddle(StoneDecoration.BLASTBRICK, 3,
 				makeIngredient(Tags.Items.INGOTS_NETHER_BRICK),
 				makeIngredient(Tags.Items.INGOTS_BRICK),
 				makeIngredient(Blocks.MAGMA_BLOCK),
-				out);
+				has(Tags.Items.INGOTS_BRICK), out);
 		addSandwich(StoneDecoration.HEMPCRETE, 6,
 				makeIngredient(IETags.clay),
 				makeIngredient(IETags.fiberHemp),
 				makeIngredient(IETags.clay),
-				out);
+				has(IETags.fiberHemp), out);
 		add3x3Conversion(StoneDecoration.COKE, IEItems.Ingredients.COAL_COKE, IETags.coalCoke, out);
 
 		addStairs(StoneDecoration.HEMPCRETE, out);
@@ -2891,11 +2892,13 @@ public class Recipes extends RecipeProvider
 		addCornerStraightMiddle(Misc.WIRE_COILS.get(WireType.COPPER_INSULATED), 4,
 				makeIngredient(IETags.fabricHemp),
 				makeIngredient(Misc.WIRE_COILS.get(WireType.COPPER)),
-				makeIngredient(IETags.fabricHemp), out);
+				makeIngredient(IETags.fabricHemp),
+				has(Misc.WIRE_COILS.get(WireType.COPPER)), out);
 		addCornerStraightMiddle(Misc.WIRE_COILS.get(WireType.ELECTRUM_INSULATED), 4,
 				makeIngredient(IETags.fabricHemp),
 				makeIngredient(Misc.WIRE_COILS.get(WireType.ELECTRUM)),
-				makeIngredient(IETags.fabricHemp), out);
+				makeIngredient(IETags.fabricHemp),
+				has(Misc.WIRE_COILS.get(WireType.ELECTRUM)), out);
 		ItemLike wireCoilRedstone = Misc.WIRE_COILS.get(WireType.REDSTONE);
 		TurnAndCopyRecipeBuilder.builder(wireCoilRedstone, 4)
 				.allowQuarterTurn()
@@ -3109,7 +3112,8 @@ public class Recipes extends RecipeProvider
 	 * @param middle the item in the middle
 	 */
 	@ParametersAreNonnullByDefault
-	private void addCornerStraightMiddle(ItemLike output, int count, Ingredient corner, Ingredient side, Ingredient middle, Consumer<FinishedRecipe> out)
+	private void addCornerStraightMiddle(ItemLike output, int count, Ingredient corner, Ingredient side, Ingredient middle,
+										 CriterionTriggerInstance condition, Consumer<FinishedRecipe> out)
 	{
 		ShapedRecipeBuilder.shaped(output, count)
 				.define('c', corner)
@@ -3118,7 +3122,7 @@ public class Recipes extends RecipeProvider
 				.pattern("csc")
 				.pattern("sms")
 				.pattern("csc")
-				.unlockedBy("has_"+toPath(output), has(output))
+				.unlockedBy("has_item", condition)
 				.save(out, toRL(toPath(output)));
 	}
 
@@ -3131,7 +3135,8 @@ public class Recipes extends RecipeProvider
 	 * @param bottom the item on the bottom
 	 */
 	@ParametersAreNonnullByDefault
-	private void addSandwich(ItemLike output, int count, Ingredient top, Ingredient middle, Ingredient bottom, Consumer<FinishedRecipe> out)
+	private void addSandwich(ItemLike output, int count, Ingredient top, Ingredient middle, Ingredient bottom,
+							 CriterionTriggerInstance condition, Consumer<FinishedRecipe> out)
 	{
 		ShapedRecipeBuilder.shaped(output, count)
 				.define('t', top)
@@ -3140,7 +3145,7 @@ public class Recipes extends RecipeProvider
 				.pattern("ttt")
 				.pattern("mmm")
 				.pattern("bbb")
-				.unlockedBy("has_"+toPath(output), has(output))
+				.unlockedBy("has_item", condition)
 				.save(out, toRL(toPath(output)));
 	}